requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-dotenv
openai
pydantic[email]
werkzeug
PyJWT
python-jose[cryptography]
